refactor netty pipeline > DisconnectHandler now handles channel inactive
