+++ src/main/java/com/hivemq/mqtt/handler/connect/ConnectHandler.java
@@ -40,7 +40,6 @@ import com.hivemq.mqtt.handler.disconnect.MqttServerDisconnector;
 import com.hivemq.mqtt.handler.ordering.OrderedTopicHandler;
 import com.hivemq.mqtt.handler.publish.DefaultPermissionsEvaluator;
 import com.hivemq.mqtt.handler.publish.FlowControlHandler;
-import com.hivemq.mqtt.message.MessageIDPools;
 import com.hivemq.mqtt.message.ProtocolVersion;
 import com.hivemq.mqtt.message.connack.CONNACK;
 import com.hivemq.mqtt.message.connect.CONNECT;
@@ -100,7 +99,6 @@ public class ConnectHandler extends SimpleChannelInboundHandler<CONNECT> impleme
     private final @NotNull PluginAuthenticatorService pluginAuthenticatorService;
     private final @NotNull PluginAuthorizerService pluginAuthorizerService;
     private final @NotNull MqttServerDisconnector mqttServerDisconnector;
-    private final @NotNull MessageIDPools messageIDPools;
 
     private int maxClientIdLength;
     private long configuredSessionExpiryInterval;
@@ -127,8 +125,7 @@ public class ConnectHandler extends SimpleChannelInboundHandler<CONNECT> impleme
             final @NotNull PluginAuthenticatorService pluginAuthenticatorService,
             final @NotNull Authorizers authorizers,
             final @NotNull PluginAuthorizerService pluginAuthorizerService,
-            final @NotNull MqttServerDisconnector mqttServerDisconnector,
-            final @NotNull MessageIDPools messageIDPools) {
+            final @NotNull MqttServerDisconnector mqttServerDisconnector) {
 
         this.onSecondConnectHandler = onSecondConnectHandler;
         this.clientSessionPersistence = clientSessionPersistence;
@@ -144,7 +141,6 @@ public class ConnectHandler extends SimpleChannelInboundHandler<CONNECT> impleme
         this.authorizers = authorizers;
         this.pluginAuthorizerService = pluginAuthorizerService;
         this.mqttServerDisconnector = mqttServerDisconnector;
-        this.messageIDPools = messageIDPools;
     }
 
     @PostConstruct
@@ -753,75 +749,6 @@ public class ConnectHandler extends SimpleChannelInboundHandler<CONNECT> impleme
                 log.trace("Client {} specified keepAlive of 0. Disabling PING mechanism", msg.getClientIdentifier());
             }
         }
-    }
-
-    @Override
-    public void channelInactive(final ChannelHandlerContext ctx) throws Exception {
-        final Channel channel = ctx.channel();
-        if (channel == null) {
-            super.channelInactive(ctx);
-            return;
-        }
-
-        final String clientId = ctx.channel().attr(ChannelAttributes.CLIENT_ID).get();
-
-        final Boolean authenticated = ctx.channel().attr(ChannelAttributes.AUTHENTICATED_OR_AUTHENTICATION_BYPASSED).get();
-        final SettableFuture<Void> disconnectFuture = ctx.channel().attr(ChannelAttributes.DISCONNECT_FUTURE).get();
-
-        //only change the session information if user is authenticated
-        if (authenticated == null || !authenticated) {
-            if (disconnectFuture != null) {
-                disconnectFuture.set(null);
-            }
-            super.channelInactive(ctx);
-            return;
-        }
-
-        final Long sessionExpiryInterval = channel.attr(ChannelAttributes.CLIENT_SESSION_EXPIRY_INTERVAL).get();
-
-        if (clientId == null || sessionExpiryInterval == null) {
-            if (disconnectFuture != null) {
-                disconnectFuture.set(null);
-            }
-            // No CONNECT message was received yet, we don't have to clean up
-            super.channelInactive(ctx);
-            return;
-        }
-
-        final boolean persistent = sessionExpiryInterval > 0;
-
-        messageIDPools.remove(clientId);
-        sendClientDisconnect(channel, clientId, persistent, sessionExpiryInterval);
-
-        super.channelInactive(ctx);
-    }
-
-    private void sendClientDisconnect(final Channel channel, final @NotNull String clientId, final boolean persistent, final long sessionExpiryInterval) {
-
-
-        final boolean preventWill = channel.attr(ChannelAttributes.PREVENT_LWT).get() != null ? channel.attr(ChannelAttributes.PREVENT_LWT).get() : false;
-        final boolean sendWill = !preventWill && (channel.attr(ChannelAttributes.SEND_WILL).get() != null ? channel.attr(ChannelAttributes.SEND_WILL).get() : true);
-        final ListenableFuture<Void> persistenceFuture = clientSessionPersistence.clientDisconnected(clientId, sendWill, sessionExpiryInterval);
-        FutureUtils.addPersistenceCallback(persistenceFuture, new FutureCallback<>() {
-                    @Override
-                    public void onSuccess(@Nullable final Void result) {
-                        if (!channel.attr(ChannelAttributes.TAKEN_OVER).get()) {
-                            channelPersistence.remove(clientId);
-                        }
-
-                        final SettableFuture<Void> disconnectFuture = channel.attr(ChannelAttributes.DISCONNECT_FUTURE).get();
-                        if (disconnectFuture != null) {
-                            disconnectFuture.set(null);
-                        }
-                    }
-
-                    @Override
-                    public void onFailure(@NotNull final Throwable throwable) {
-                        Exceptions.rethrowError("Unable to update client session data for disconnecting client " + clientId +
-                                " with clean session set to " + !persistent + ".", throwable);
-                    }
-                }
-        );
     }
 
     private double getGracePeriod() {
